Relocate Plugin-related features to loki_bms_common.plugins namespace - Makes more sense than having them floating around in the root folder
