+++ Extensions/Extensions.cs
@@ -1,4 +1,5 @@
-﻿using SkiaSharp;
+﻿using loki_bms_common.Database;
+using SkiaSharp;
 using System;
 using System.Collections.Generic;
 using System.Linq;
