+++ loki-bms-common/Plugins/LokiCustomMenu.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace loki_bms_common
+namespace loki_bms_common.Plugins
 {
     public abstract class LokiCustomMenu
     {
-        public string XamlSourcePath = string.Empty;
+        public string XamlFilePath = string.Empty;
         public object? GeneratedElements;
         public string ButtonText = "???";
     }
