+++ loki-bms-common/Plugins/TestPlugin.cs
@@ -4,10 +4,10 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using loki_bms_common;
+using loki_bms_common.Plugins;
 
 [assembly: LokiPlugin(typeof(TestPlugin), "Test Plugin", "999.0.0", Author = "Dani Lodholm")]
-namespace loki_bms_common
+namespace loki_bms_common.Plugins
 {
     public class TestPlugin : LokiPlugin
     {
